Merge in PE_RBVD_APP-ID-105529_DSG/dto_enterprise_life_insurance from bugfix/modificationInTypeOfEmployeesPayrollAmount to release/0.1

* commit 'f6531df2003b69a25c327d39073b0d4c37dc723d':
  feat:PERAR-26902 PERAR-27007 Modification in Type of PayrollEmployeeNumber

+++ src/main/java/com/bbva/rbvd/dto/enterpriseinsurance/modifyquotation/dao/InsuranceQuoteCoLifeDAO.java
@@ -8,7 +8,7 @@ public class InsuranceQuoteCoLifeDAO {
     private String insuranceModalityType;
     private BigDecimal incomesPayrollAmount;
     private String currencyId;
-    private int payrollEmployeeNumber;
+    private BigDecimal payrollEmployeeNumber;
     private String employeeEmailName;
     private int yearsOldEmployeesIndType;
     private String creationUserId;
@@ -56,11 +56,11 @@ public class InsuranceQuoteCoLifeDAO {
         this.currencyId = currencyId;
     }
 
-    public int getPayrollEmployeeNumber() {
+    public BigDecimal getPayrollEmployeeNumber() {
         return payrollEmployeeNumber;
     }
 
-    public void setPayrollEmployeeNumber(int payrollEmployeeNumber) {
+    public void setPayrollEmployeeNumber(BigDecimal payrollEmployeeNumber) {
         this.payrollEmployeeNumber = payrollEmployeeNumber;
     }
 
@@ -148,7 +148,7 @@ public class InsuranceQuoteCoLifeDAO {
             return Builder.this;
         }
 
-        public Builder withPayrollEmployeeNumber(int payrollEmployeeNumber) {
+        public Builder withPayrollEmployeeNumber(BigDecimal payrollEmployeeNumber) {
             insuranceQuoteCoLifeDAO.payrollEmployeeNumber = payrollEmployeeNumber;
             return Builder.this;
         }
